minimize processing before sending MIDI notes
Make it so that column indicates what to play next, not what was just
played. This allows noteOff and noteOn to be sent immediately when
tick() is entered. Timing is everything.

Most significant downside of this approach is that we have to copy
activeNote and activeChannel when switching pattern. Ugly.